# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoonPlants API",
    "version": "1.0.0",
    "description": "REST API for MoonPlants — smart plant monitoring and automated watering system. Consists of two namespaces:\n\n- **`/api/v1/`** — User-facing endpoints (Supabase JWT auth)\n- **`/api/iot/v1/`** — Device-facing endpoints (HMAC-SHA256 auth)",
    "contact": {
      "name": "MoonPlants"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://your-production-domain.com",
      "description": "Production (replace with actual domain)"
    }
  ],
...
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
      "Device": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "display_name": {
            "type": "string",
            "example": "Balcony Hub #1"
          },
          "channels_count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 16,
            "example": 4
          },
          "supports_pumps": {
            "type": "boolean"
          },
          "supports_light": {
            "type": "boolean"
          },
          "firmware_version": {
            "type": [
              "string",
              "null"
            ],
            "example": "1.2.0"
          },
          "status": {
            "type": "string",
            "enum": [
              "unclaimed",
              "claimed",
              "revoked"
            ],
            "example": "claimed"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeviceDetail": {
        "type": "object",
        "properties": {
          "device": {
            "$ref": "#/components/schemas/Device"
          },
          "plants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantSummary"
            }
          },
          "pendingCommands": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceCommand"
            }
          }
        }
      },
      "DeviceCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "PUMP_WATER",
              "LIGHT_ON",
              "LIGHT_OFF"
            ],
            "example": "PUMP_WATER"
          },
          "payload": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "sent",
              "acked",
              "failed",
              "expired"
            ],
            "example": "queued"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "send_after": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "PlantSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Monstera Deliciosa"
          },
          "species_name": {
            "type": "string"
          },
          "soil_channel": {
            "type": [
              "integer",
              "null"
            ]
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "auto_watering_enabled": {
            "type": "boolean"
          },
          "auto_light_enabled": {
            "type": "boolean"
          }
        }
      },
...
      "LightCommandRequest": {
        "type": "object",
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "on",
              "off",
              "on_for"
            ]
          },
          "durationSec": {
            "type": "integer",
            "minimum": 1,
            "maximum": 86400,
            "description": "Required when mode = on_for"
          }
        }
      },
...
      "PerenualResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Monstera deliciosa (Swiss cheese plant)"
          },
          "image": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT access token. Pass via Authorization: Bearer <token>"
      },
      "IotHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Signature",
        "description": "HMAC-SHA256 device signature. IoT endpoints require all 5 headers: X-Device-Id, X-Device-Seq, X-Device-Timestamp, X-Content-SHA256, X-Device-Signature. Canonical string = METHOD\\nPATH\\nX-Device-Id\\nX-Device-Seq\\nX-Device-Timestamp\\nX-Content-SHA256"
      }
    }
  },
  "paths": {
...
    "/api/v1/devices/{deviceId}": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "Get device details",
        "description": "Returns device info, linked plants, and pending commands.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                },
                "examples": {
                  "success": {
                    "summary": "Device with one plant",
                    "value": {
                      "device": {
                        "id": "e2a3b1c4-0000-0000-0000-000000000001",
                        "display_name": "Balcony Hub",
                        "channels_count": 4,
                        "supports_pumps": true,
                        "supports_light": true,
                        "firmware_version": "1.2.0",
                        "status": "claimed",
                        "last_seen_at": "2026-03-03T10:00:00Z",
                        "created_at": "2026-01-01T00:00:00Z"
                      },
                      "plants": [
                        {
                          "id": "plant-uuid-001",
                          "name": "Monstera",
                          "species_name": "Monstera deliciosa",
                          "soil_channel": 1,
                          "image_url": null,
                          "auto_watering_enabled": false,
                          "auto_light_enabled": false
                        }
                      ],
                      "pendingCommands": []
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/devices/{deviceId}/actions/light": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Send light command",
        "description": "Queues a LIGHT_ON or LIGHT_OFF command for the device. The device picks it up on next poll.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LightCommandRequest"
              },
              "examples": {
                "turnOn": {
                  "summary": "Turn light on for 2 hours",
                  "value": {
                    "mode": "on_for",
                    "durationSec": 7200
                  }
                },
                "turnOff": {
                  "summary": "Turn light off",
                  "value": {
                    "mode": "off"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate command (deduplicated)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "deduplicated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Command queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "commandId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or device does not support light",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Active light command already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/perenual-autocomplete": {
      "get": {
        "tags": [
          "Utility"
        ],
        "summary": "Autocomplete plant species",
        "description": "Searches the Perenual API for plant species by name. Returns a list of matches with ID, display name, and thumbnail image URL.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "description": "Search query string",
            "example": "monstera"
          }
        ],
        "responses": {
          "200": {
            "description": "Autocomplete results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PerenualResult"
                      }
                    }
                  }
                },
                "examples": {
                  "monstera": {
                    "summary": "Results for 'monstera'",
                    "value": {
                      "data": [
                        {
                          "id": 1882,
                          "name": "Monstera deliciosa (Swiss cheese plant)",
                          "image": "https://perenual.com/storage/species_img/1882/small/1882.jpg"
                        }
                      ]
                    }
                  },
                  "empty": {
                    "summary": "No query – empty result",
                    "value": {
                      "data": []
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}